+++ mkpage_test.go
@@ -20,7 +20,6 @@ package mkpage
 
 import (
 	"bytes"
-	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"path"
@@ -73,7 +72,7 @@ func TestResolveData(t *testing.T) {
 		t.Error(err)
 		t.FailNow()
 	}
-	buf, err := gomarkdownProcessor(src)
+	buf, err := pandocProcessor(src, "markdown", "html")
 	expected := string(buf)
 
 	if err := checkMap("Nav", expected, data); err != nil {
@@ -86,7 +85,7 @@ func TestResolveData(t *testing.T) {
 		t.Error(err)
 		t.FailNow()
 	}
-	buf, err = gomarkdownProcessor(src)
+	buf, err = pandocProcessor(src, "markdown", "html")
 	expected = string(buf)
 
 	if err := checkMap("Content", expected, data); err != nil {
@@ -136,69 +135,6 @@ Weather: {{.weather.data.text}}
 	}
 	checkForString(out, "Hi there!")
 	checkForString(out, "<ul>")
-}
-
-func TestBasic(t *testing.T) {
-	src := `
-
-# Opening Slide
-
-## This is an introduction
-
-+ me, <me@example.org>
-+ [mygitrepo.org](http://mygitrepo.org)
-+ February 31, 2016, Somebody's University, Someplace
-
----
-
-# Slide Two
-
-| col A | col B | Col C |
-|-------|-------|-------|
-| A     | One   | 1     |
-
----
-
-## Slide Three
-
-This is slide three, just a random paragraph of text. Blah, blah, blah, blah, blah, blah, boink!
-
-
-`
-
-	tmpl, err := template.New("slides.tmpl").Parse(DefaultSlideTemplateSource)
-	if err != nil {
-		t.Errorf("Can't parse DefaultTemplateSource templates %s", err)
-		t.FailNow()
-	}
-
-	titles := []string{
-		"<h1>Opening Slide</h1>",
-		"<h1>Slide Two</h1>",
-		"<h2>Slide Three</h2>",
-	}
-
-	slides, err := MarkdownToSlides("test.html", []byte(src))
-	if err != nil {
-		t.Errorf("MarkdownToSlides(%q, ...) error %s", "test.html", err)
-		t.FailNow()
-	}
-	if len(slides) != 3 {
-		src, _ := json.Marshal(slides)
-		t.Errorf("Was expected three slides %+v\n%s", slides, src)
-	}
-
-	keyVals := map[string]string{}
-	for i, slide := range slides {
-		keyVals["Title"] = "text:" + titles[i]
-		s, err := MakeSlideString("slides.tmpl", tmpl, keyVals, slide)
-		if err != nil {
-			t.Errorf("MakeSlideString() failed %d - %s", i, err)
-		}
-		if strings.Contains(s, titles[i]) == false {
-			t.Errorf("Expected %q in slide %d -> %s", titles[i], i, s)
-		}
-	}
 }
 
 func TestGrep(t *testing.T) {
@@ -276,14 +212,14 @@ This is text **in bold** in text.
 	srcCRLF := bytes.Replace(srcRaw, []byte("\n"), []byte("\r\n"), -1)
 	srcLF := srcRaw
 	// Render HTML using normalize Unix eol
-	src1, err := gomarkdownProcessor(srcLF)
+	src1, err := pandocProcessor(srcLF, "markdown", "html")
 	if err != nil {
-		t.Errorf("gomarkdownProcessor(srcCRLF) error %s", err)
+		t.Errorf("pandocProcessor(srcLF, 'markdown', 'html') error %s", err)
 	}
 	// Render HTML using normalize old DOS eol
-	src2, err := gomarkdownProcessor(srcCRLF)
+	src2, err := pandocProcessor(srcCRLF, "markdown", "html")
 	if err != nil {
-		t.Errorf("gomarkdownProcessor(srcLF) error %s", err)
+		t.Errorf("pandocProcessor(srcCRLF, 'markdown', 'html') error %s", err)
 	}
 	if bytes.Compare(src1, src2) != 0 {
 		t.Errorf("expected (eol normalized) ->\n%s\ngot ->\n%s\n",